+++ tests/ca_capacity_test.py
@@ -12,12 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from grscheller.circular_array import CircularArray
+from __future__ import annotations
+
+from typing import Optional
+from grscheller.circular_array.ca import CircularArray
 
 class TestCapacity:
 
     def test_capacity_original(self) -> None:
-        c = CircularArray()
+        c: CircularArray[object] = CircularArray()
         assert c.capacity() == 2
 
         c = CircularArray(1, 2)
@@ -54,7 +57,7 @@ class TestCapacity:
         assert c.fractionFilled() == 2/3
 
     def test_double(self) -> None:
-        c = CircularArray(1, 2, 3)
+        c: CircularArray[Optional[int]] = CircularArray(1, 2, 3)
         assert c.popL() == 1
         assert c.capacity() == 3
         c.double()
@@ -85,7 +88,7 @@ class TestCapacity:
             jj -= 1
 
     def test_empty(self) -> None:
-        c = CircularArray()
+        c: CircularArray[object] = CircularArray()
         assert c == CircularArray()
         assert c.capacity() == 2
         c.double()